Working issues 6

+++ app/src/main/java/com/example/aas_app/data/dao/DemoTemplatesDao.kt
@@ -22,5 +22,5 @@ interface DemoTemplatesDao {
     suspend fun getAllDemoTemplates(): List<DemoTemplatesEntity>
 
     @Query("SELECT * FROM demotemplates WHERE id = :id")
-    suspend fun getTemplateById(id: Int): DemoTemplatesEntity?
+    suspend fun getTemplateById(id: Long): DemoTemplatesEntity?
 }
