+++ app/src/main/java/com/example/aas_app/data/entity/InstructorStudentAssignmentEntity.kt
@@ -5,8 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "instructor_student_assignments")
 data class InstructorStudentAssignmentEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val instructor_id: Int,
-    val student_id: Int,
-    val program_id: Int? = null
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    val instructor_id: Long,
+    val student_id: Long,
+    val program_id: Long? = null
 )
