+++ app/src/main/java/com/example/aas_app/data/entity/PeclEvaluationResultEntity.kt
@@ -5,10 +5,10 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "pecl_evaluation_results")
 data class PeclEvaluationResultEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val student_id: Int,
-    val instructor_id: Int,
-    val question_id: Int,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    val student_id: Long,
+    val instructor_id: Long,
+    val question_id: Long,
     val score: Double,
     val comment: String = "",
     val timestamp: Long
