+++ app/src/main/java/com/example/aas_app/data/entity/PeclPoiEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "pecl_pois")
 data class PeclPoiEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
-    val program_id: Int
+    val program_id: Long
 )
