+++ app/src/main/java/com/example/aas_app/data/entity/PeclTaskEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "pecl_tasks")
 data class PeclTaskEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val name: String,
-    val poi_id: Int
+    val poi_id: Long
 )
