+++ app/src/main/java/com/example/aas_app/data/entity/QuestionAssignmentEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "question_assignments")
 data class QuestionAssignmentEntity(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val question_id: Int,
-    val task_id: Int
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    val question_id: Long,
+    val task_id: Long
 )
